refactor(planning): move ObstacleConfig into per-robot specs
Give taller platforms like B2/G1 their own collision height bands instead of a shared planning-node default.
